app/requirements.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
djangorestframework
djangorestframework-jwt
django-cors-headers
django-environ
python-dotenv
numpy==1.21.2
matplotlib==3.7.1
pandas
psycopg2
# statsmodels==0.12.2
scikit-learn==1.0
surprise
mlxtend
pymongo